requirements.txt
chainlit>=0.7.0
neo4j>=5.8.0
python-dotenv>=1.0.0
litellm>=1.0.0,<2.0.0
#PyPDF2>=3.0.0
tenacity>=8.2.0
PyYAML>=6.0
pandas>=2.0.0
numpy>=1.24.0
asyncio>=3.4.3
pdfplumber>=0.10.0
pymupdf>=1.24.0
sentence-transformers>=2.2.2
qdrant-client>=1.6
fastembed>=0.2
hf_xet
nltk>=3.8.1
presidio-analyzer>=2.2.0
mcp
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
httpx>=0.24.0